requirements.txt
black==25.12.0
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.1
distlib==0.4.0
filelock==3.20.3
identify==2.6.16
idna==3.11
iniconfig==2.3.0
isort==7.0.0
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
requests==2.32.5
ruff==0.14.11
urllib3==2.6.3
virtualenv==20.36.1
